requirements.txt
openai>=1.30.0
langchain-openai>=0.1.0
azure-identity>=1.15.0
streamlit>=1.35.0
python-dotenv>=1.0.0
requests>=2.31.0
